# alx_api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ALX API
  version: 0.1.0
  description: >-
    Auth Core with local PostgreSQL integration coverage and least-privilege repositories.
    Legacy credentials are verified inside a restricted database function and upgraded to Argon2id on first successful login when enabled.
    Real production-user credentials have not been bulk-migrated; production password changes remain disabled until client cutover and delivery configuration.
servers:
  - url: http://127.0.0.1:3001
    description: Local development only
paths:
...
  /api/v1/users:
    get:
      operationId: listUsers
      summary: List system users
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    post:
      operationId: createUser
      summary: Create a system user
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
  /api/v1/users/{id}:
    get:
      operationId: getUser
      summary: Get user details
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    patch:
      operationId: updateUser
      summary: Update user details or disable status
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
  /api/v1/users/{id}/roles:
    get:
      operationId: getUserRoles
      summary: Get user assigned roles
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    post:
      operationId: setUserRoles
      summary: Assign or update user roles
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data, requestId]
      properties:
        success: { type: boolean, const: true }
        data: { type: object }
        meta:
          type: object
          additionalProperties: true
        requestId: { type: string }
...